% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HadesExtras.R
\name{connectionHandlerFromList}
\alias{connectionHandlerFromList}
\title{Create a ConnectionHandler from a configuration list}
\usage{
connectionHandlerFromList(configConnection)
}
\arguments{
\item{configConnection}{A list containing connection configuration settings:
\itemize{
\item tempEmulationSchema: Schema to use for temp table emulation
\item connectionDetailsSettings: Settings for creating connection details
}}
}
\value{
A ConnectionHandler object
}
\description{
Create a ConnectionHandler from a configuration list
}
