% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAtomicCodeCountsTable.R
\name{createAtomicCodeCountsTable}
\alias{createAtomicCodeCountsTable}
\title{Create code counts table}
\usage{
createAtomicCodeCountsTable(
  CDMdbHandler,
  domains = NULL,
  codeAtomicCountsTable = "code_atomic_counts"
)
}
\arguments{
\item{CDMdbHandler}{A CDMdbHandler object that contains database connection details}
}
\value{
Nothing. Creates a table called 'code_counts' in the results schema with columns:
\itemize{
\item domain - The domain of the code (Condition, Drug, etc.)
\item concept_id - The OMOP concept ID
\item calendar_year - The year of the events
\item gender_concept_id - The gender concept ID
\item age_decile - The age decile (0-9, 10-19, etc.)
\item event_counts - Number of events for this code
\item person_counts - Number of persons with this code
\item incidence_person_counts - Number of persons with first occurrence of this code
\item descendant_event_counts - Number of events including descendant concepts
\item descendant_person_counts - Number of persons including descendant concepts
\item descendant_incidence_person_counts - Number of persons with first occurrence including descendants
\item total_person_counts - Total number of persons in the stratum
}
}
\description{
Creates a table containing counts of codes by domain, concept, year, gender and age decile.
The table includes both event counts and person counts, as well as descendant counts.
}
