% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotingFunctions.R
\name{createCodeCountsTableFromResults}
\alias{createCodeCountsTableFromResults}
\title{Create a reactable table from results}
\usage{
createCodeCountsTableFromResults(results)
}
\arguments{
\item{results}{List containing concept relationships and counts data}
}
\value{
A reactable table object
}
\description{
Create a reactable table from results
}
