% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotingFunctions.R
\name{createMermaidGraphFromResults}
\alias{createMermaidGraphFromResults}
\title{Create a Mermaid graph visualization from results}
\usage{
createMermaidGraphFromResults(results, showsMappings = FALSE)
}
\arguments{
\item{results}{List containing concept relationships and counts data}

\item{showsMappings}{Logical indicating whether to show concept mappings}
}
\value{
String containing Mermaid graph definition
}
\description{
Create a Mermaid graph visualization from results
}
