% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotingFunctions.R
\name{createPlotFromResults}
\alias{createPlotFromResults}
\title{Create a plotly visualization from results}
\usage{
createPlotFromResults(results, showsMappings = FALSE, ...)
}
\arguments{
\item{results}{List containing concept relationships and counts data}

\item{showsMappings}{Logical indicating whether to show concept mappings}

\item{...}{Additional arguments passed to plotly::ggplotly}
}
\value{
A plotly object
}
\description{
Create a plotly visualization from results
}
