% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotingFunctions.R
\name{.conceptIdToRGB}
\alias{.conceptIdToRGB}
\title{Convert concept ID to RGB color}
\usage{
.conceptIdToRGB(conceptId)
}
\arguments{
\item{conceptId}{Numeric concept ID}
}
\value{
Character string with hex color code
}
\description{
Convert concept ID to RGB color
}
\keyword{internal}
