% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_logs.R
\name{.createConsoleAppenderForSandboxLogging}
\alias{.createConsoleAppenderForSandboxLogging}
\title{Create Console Appender for Sandbox Logging}
\usage{
.createConsoleAppenderForSandboxLogging(
  layout = ParallelLogger::layoutParallel,
  sToken
)
}
\arguments{
\item{layout}{A layout function for the logger. Defaults to ParallelLogger::layoutParallel.}

\item{sToken}{The session token to be added to log entries}
}
\value{
An appender object for logging.
}
\description{
Creates a console appender for sandbox logging with a specified layout.
}
