% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotingFunctions.R
\name{.genderConceptIdToLabel}
\alias{.genderConceptIdToLabel}
\title{Convert a gender concept ID to a display label}
\usage{
.genderConceptIdToLabel(genderConceptId)
}
\arguments{
\item{genderConceptId}{Integer OMOP gender concept ID (8507 male, 8532 female)}
}
\value{
Character label
}
\description{
Convert a gender concept ID to a display label
}
\keyword{internal}
