% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAPIInfo.R
\name{getAPIInfo}
\alias{getAPIInfo}
\title{Get the API information}
\usage{
getAPIInfo(CDMdbHandler)
}
\arguments{
\item{CDMdbHandler}{A CDMdbHandler object that contains database connection details}
}
\value{
A tibble containing the CDM source information with columns from the cdm_source table
}
\description{
Retrieves the CDM source information from the vocabulary schema of an OMOP CDM database.
Retrieves the API information from the vocabulary schema of an OMOP CDM database.
This includes metadata about the API, version, and other administrative details.
}
