% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataFunctions.R
\name{getCodeCounts}
\alias{getCodeCounts}
\title{Get code counts and related concept information}
\usage{
getCodeCounts(CDMdbHandler, conceptIds)
}
\arguments{
\item{CDMdbHandler}{A CDMdbHandler object that contains database connection details}

\item{conceptIds}{Vector of concept IDs to get counts and relationships for}
}
\value{
A list containing:
\itemize{
\item concept_relationships - Tibble of concept relationships (Maps to, Subsumes)
\item concepts - Tibble of concept details for related concepts
\item code_counts - Tibble of code counts from the code_counts table
}
}
\description{
Get code counts and related concept information
}
