% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConceptsWithCodeCounts.R
\name{getConceptsWithCodeCounts}
\alias{getConceptsWithCodeCounts}
\title{Get list of concepts with code counts}
\usage{
getConceptsWithCodeCounts(CDMdbHandler, codeCountsTable = "code_counts")
}
\arguments{
\item{CDMdbHandler}{A CDMdbHandler object that contains database connection details}

\item{codeCountsTable}{Name of the code counts table in the results schema. Defaults to "code_counts"}
}
\value{
A tibble with columns:
\itemize{
\item \code{concept_id} - The OMOP concept ID
\item \code{concept_name} - The human-readable concept name
\item \code{domain_id} - The domain identifier (e.g., Condition, Procedure, Drug, etc.)
\item \code{vocabulary_id} - The vocabulary identifier (e.g., SNOMED, ICD10)
\item \code{concept_class_id} - The concept class
\item \code{standard_concept} - Logical indicating if this is a standard concept
\item \code{concept_code} - The concept code
\item \code{record_counts} - Number of events for this code
\item \code{descendant_record_counts} - Number of events including descendant concepts
\item \code{number_of_descendants} - Number of descendant concepts (including itself)
\item \code{person_counts} - Number of distinct persons with this code
\item \code{descendant_person_counts} - Number of distinct persons including descendant concepts
}
}
\description{
Retrieves a list of concepts that have associated code counts in the results schema.
This function provides concept metadata including names, vocabulary IDs, and standard concept flags
for concepts that are actively used in the database.
}
