% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConceptTree.R
\name{getConceptTree_memoise}
\alias{getConceptTree_memoise}
\title{Memoised version of getConceptTree}
\usage{
getConceptTree_memoise(
  CDMdbHandler,
  conceptId,
  codeCountsTable = "code_counts"
)
}
\arguments{
\item{CDMdbHandler}{A CDMdbHandler object that contains database connection details}

\item{conceptId}{The concept ID to get the tree for}

\item{codeCountsTable}{Name of the code counts table in the results schema. Defaults to "code_counts"}
}
\value{
Same shape as \code{\link{getConceptTree}}.
}
\description{
A memoised version of the getConceptTree function that caches results to improve performance
for repeated calls with the same parameters. The CDMdbHandler argument is omitted from
the cache key to allow sharing across different database connections.
}
