% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataFunctions.R
\name{getListOfConcepts}
\alias{getListOfConcepts}
\title{Get list of concepts with code counts}
\usage{
getListOfConcepts(CDMdbHandler)
}
\arguments{
\item{CDMdbHandler}{A CDMdbHandler object that contains database connection details}
}
\value{
A tibble with concept_id, concept_name, vocabulary_id, and standard_concept for concepts present in code_counts
}
\description{
Get list of concepts with code counts
}
