% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLogs.R
\name{getLogs}
\alias{getLogs}
\title{Get logs from ParallelLogger temp file}
\usage{
getLogs()
}
\value{
A character vector with each log line as a string.
}
\description{
Retrieves log messages from the ParallelLogger temporary log file.
This may help track internal API events, errors, and debug information.
}
