% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HadesExtras.R
\name{HadesExtras_createCDMdbHandlerFromList}
\alias{HadesExtras_createCDMdbHandlerFromList}
\title{createCDMdbHandlerFromList}
\usage{
HadesExtras_createCDMdbHandlerFromList(
  config,
  loadConnectionChecksLevel = "allChecks"
)
}
\arguments{
\item{config}{A list containing configuration settings for the CDMdbHandler.
\itemize{
\item databaseName: The name of the database.
\item connection: A list of connection details settings.
\item cdm: A list of CDM database schema settings.
\item cohortTable: The name of the cohort table.
}}

\item{loadConnectionChecksLevel}{The level of checks to perform when loading the connection.}
}
\value{
A CDMdbHandler object.
}
\description{
A function to create a CDMdbHandler object from a list of configuration settings.
}
