% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendFeedback.R
\name{sendFeedback}
\alias{sendFeedback}
\title{Send feedback to the API server}
\usage{
sendFeedback(feedback)
}
\arguments{
\item{feedback}{A character string containing the feedback to send.}
}
\description{
Send feedback to the API server
}
