% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotingFunctions.R
\name{createAgeHistogramFromPersonCounts}
\alias{createAgeHistogramFromPersonCounts}
\title{Create a bar chart of person counts by age decile}
\usage{
createAgeHistogramFromPersonCounts(filterPersonCounts)
}
\arguments{
\item{filterPersonCounts}{The `filter_person_counts` tibble from \code{\link{getPersonCounts}}}
}
\value{
A plotly object
}
\description{
Create a bar chart of person counts by age decile
}
