% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotingFunctions.R
\name{createSexPieChartFromPersonCounts}
\alias{createSexPieChartFromPersonCounts}
\title{Create a pie chart of person counts by sex}
\usage{
createSexPieChartFromPersonCounts(filterPersonCounts)
}
\arguments{
\item{filterPersonCounts}{The `filter_person_counts` tibble from \code{\link{getPersonCounts}}}
}
\value{
A plotly object
}
\description{
Create a pie chart of person counts by sex
}
