% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataFunctions.R
\name{getCDMSource}
\alias{getCDMSource}
\title{Get the CDM source information}
\usage{
getCDMSource(CDMdbHandler)
}
\arguments{
\item{CDMdbHandler}{A CDMdbHandler object that contains database connection details}
}
\value{
A tibble of the CDM source information
}
\description{
Get the CDM source information
}
