% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConceptsWithCodeCounts.R
\name{getConceptsWithCodeCounts}
\alias{getConceptsWithCodeCounts}
\title{Get list of concepts with code counts}
\usage{
getConceptsWithCodeCounts(CDMdbHandler, codeCountsTable = "code_counts")
}
\arguments{
\item{CDMdbHandler}{A CDMdbHandler object that contains database connection details}

\item{codeCountsTable}{Name of the code counts table in the results schema. Defaults to "code_counts"}
}
\value{
A tibble with columns:
\itemize{
\item \code{concept_id} - The OMOP concept ID
\item \code{concept_name} - The human-readable concept name
\item \code{vocabulary_id} - The vocabulary identifier (e.g., SNOMED, ICD10)
\item \code{standard_concept} - Logical indicating if this is a standard concept
}
}
\description{
Retrieves a list of concepts that have associated code counts in the results schema.
This function provides concept metadata including names, vocabulary IDs, and standard concept flags
for concepts that are actively used in the database.
}
