% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotingFunctions.R
\name{pruneLevelsFromResults}
\alias{pruneLevelsFromResults}
\title{Prune levels from results}
\usage{
pruneLevelsFromResults(results, pruneLevels, pruneClass = NULL)
}
\arguments{
\item{results}{A list of results from getCodeCounts}

\item{pruneLevels}{The levels to prune}
}
\value{
A list of results with the levels pruned
}
\description{
Prune levels from results
}
