% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApiServer.R
\name{runApiServer}
\alias{runApiServer}
\title{Run the ROMOPAPI API server}
\usage{
runApiServer(
  cohortTableHandlerConfig = NULL,
  host = "127.0.0.1",
  port = 8585,
  ...
)
}
\arguments{
\item{cohortTableHandlerConfig}{Configuration for connecting to the OMOP CDM database. If NULL, uses test Eunomia database}

\item{host}{Host address to run the API server on. Defaults to "127.0.0.1"}

\item{port}{Port number to run the API server on. Defaults to 8585}

\item{...}{Additional arguments passed to plumber::pr_run()}
}
\description{
Run the ROMOPAPI API server
}
